cleaning up and adding doc strings
