+++ src/chains/near.ts
@@ -7,6 +7,14 @@ export const ONE_YOCTO = "1";
 
 type NetworkId = "mainnet" | "testnet";
 
+/**
+ * Extracts the network ID from a given NEAR account ID.
+ * If the account ID does not end with "near" or "testnet", it logs a warning.
+ * Defaults to "mainnet" if the network ID is not "testnet".
+ *
+ * @param accountId - The NEAR account ID to extract the network ID from.
+ * @returns The network ID, either "mainnet" or "testnet".
+ */
 export function getNetworkId(accountId: string): NetworkId {
   const accountExt = accountId.split(".").pop() || "";
   if (!["near", "testnet"].includes(accountExt)) {
@@ -18,6 +26,12 @@ export function getNetworkId(accountId: string): NetworkId {
   return accountExt !== "testnet" ? "mainnet" : accountExt;
 }
 
+/**
+ * Generates a NEAR configuration object based on the provided network ID.
+ *
+ * @param networkId - The network ID, either "mainnet" or "testnet".
+ * @returns A NearConfig object containing the network ID and node URL.
+ */
 export function configFromNetworkId(networkId: NetworkId): NearConfig {
   return {
     networkId,
@@ -27,11 +41,11 @@ export function configFromNetworkId(networkId: NetworkId): NearConfig {
 
 /**
  * Loads Near Account from provided keyPair and accountId
- * Defaults to TESTNET_CONFIG
+ *
  * @param keyPair {KeyPair}
  * @param accountId {string}
  * @param network {NearConfig} network settings
- * @returns {Account}
+ * @returns A Promise that resolves to a NEAR Account instance.
  */
 export const nearAccountFromKeyPair = async (config: {
   keyPair: KeyPair;
@@ -41,8 +55,12 @@ export const nearAccountFromKeyPair = async (config: {
   return createNearAccount(config.accountId, config.network, config.keyPair);
 };
 
-/** Minimally sufficient Account instance to construct the signing contract instance.
+/** Minimally sufficient Account instance to construct readonly MpcContract connection.
  *  Can't be used to change methods.
+ *
+ * @param accountId {string}
+ * @param network {NearConfig} network settings
+ * @returns A Promise that resolves to a NEAR Account instance.
  */
 export const nearAccountFromAccountId = async (
   accountId: string,
@@ -51,6 +69,14 @@ export const nearAccountFromAccountId = async (
   return createNearAccount(accountId, network);
 };
 
+/**
+ * Creates a NEAR account instance using the provided account ID, network configuration, and optional key pair.
+ *
+ * @param accountId - The NEAR account ID.
+ * @param network - The NEAR network configuration.
+ * @param keyPair - (Optional) The key pair for the account.
+ * @returns A Promise that resolves to a NEAR Account instance.
+ */
 export const createNearAccount = async (
   accountId: string,
   network: NearConfig,
